# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Mockup Test API Docs",
        "description": "This is API Documentation for Mockup Test FSW",
        "termsOfService": "http://localhost:5000/tos",
        "contact": {
            "email": "ftaufik.ros@gmail.com",
            "name": "Taufik Firmansyah"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit/"
        }
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/todo/delete/{id}": {
            "delete": {
                "description": "delete todo task",
                "operationId": "deleteTodo",
                "tags": [
                    "todo"
                ],
                "parameters": [
                    {
                        "required": true,
                        "name": "id",
                        "in": "path",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success deleting data",
                        "schema": {
                            "$ref": "#/definitions/successApiResponseMultiple"
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "schema": {
                            "$ref": "#/definitions/errorApiResponseSingle"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorApiResponseSingle"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "successApiResponseMultiple": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "data": {
                    "type": "array"
                }
            }
        },
        "errorApiResponseSingle": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {}
}